requirements.txt
Flask==2.2.2
Flask-Reuploaded==1.2.0
Flask-WTF==1.0.1
pillow==9.2.0
numpy==1.21.5
scipy==1.7.3
scikit-learn==1.1.1